area controller levédve authenticatorral és a usercontroller is elkezdve

+++ controller/src/main/java/hu/otemplom/karbantarto/controller/AreaController.java
@@ -1,48 +1,71 @@
 package hu.otemplom.karbantarto.controller;
 
 
+import hu.otemplom.karbantarto.controller.exceptions.InvalidTokenException;
 import hu.otemplom.karbantarto.model.Area;
 import hu.otemplom.karbantarto.model.Exceptions.Area.InvalidIdException;
+import hu.otemplom.karbantarto.model.Exceptions.User.InvalidFullNameException;
+import hu.otemplom.karbantarto.model.Exceptions.User.InvalidRoleException;
+import hu.otemplom.karbantarto.model.Exceptions.User.InvalidUsernameException;
 import hu.otemplom.karbantarto.service.AreaService;
 import hu.otemplom.karbantarto.service.Exceptions.AreaService.AreaAlreadyExistsException;
 import hu.otemplom.karbantarto.service.Exceptions.AreaService.AreaDoesNotExistsException;
 import hu.otemplom.karbantarto.service.Exceptions.AreaService.InvalidAreaException;
-import hu.otemplom.karbantarto.service.impl.AreaServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-
-import javax.websocket.server.PathParam;
+import org.springframework.web.server.ResponseStatusException;
 import java.util.Collection;
-import java.util.List;
 @RequestMapping("api/v1/area")
 @RestController
 public class AreaController {
     @Autowired
-    public AreaController(AreaService areaService) {
+    public AreaController(AreaService areaService, IUserAuthenticator authenticator) {
+
         this.areaService = areaService;
+        this.authenticator =  authenticator;
     }
 
     AreaService areaService;
+    IUserAuthenticator authenticator;
     @GetMapping
-    public Collection<Area> getAllArea(){
-        return areaService.getAllArea();
+    public Collection<Area> getAllArea(@RequestHeader("authorization") String rawToken) throws hu.otemplom.karbantarto.model.Exceptions.User.InvalidIdException, InvalidRoleException, InvalidFullNameException, InvalidUsernameException, InvalidTokenException {
+        if(authenticator.userIsAdmin(rawToken)){
+            return areaService.getAllArea();
+        }
+        else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"",new InvalidTokenException("meow"));
+
     }
     @PostMapping
-    public void addArea(@RequestBody Area area) throws InvalidIdException, InvalidAreaException, AreaAlreadyExistsException {
-        areaService.addArea(area);
+    public void addArea(@RequestHeader("authorization") String rawToken, @RequestBody Area area) throws InvalidIdException, InvalidAreaException, AreaAlreadyExistsException, hu.otemplom.karbantarto.model.Exceptions.User.InvalidIdException, InvalidRoleException, InvalidFullNameException, InvalidUsernameException, InvalidTokenException {
+        if(authenticator.userIsAdmin(rawToken)){
+            areaService.addArea(area);
+        }
+        else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"",new InvalidTokenException("meow"));
     }
     @PutMapping
-    public void modifyArea(@RequestBody Area area) throws InvalidAreaException, AreaDoesNotExistsException {
-        areaService.modifyArea(area);
+    public void modifyArea(@RequestHeader("authorization") String rawToken,@RequestBody Area area) throws InvalidAreaException, AreaDoesNotExistsException, hu.otemplom.karbantarto.model.Exceptions.User.InvalidIdException, InvalidRoleException, InvalidFullNameException, InvalidUsernameException, InvalidTokenException {
+        if(authenticator.userIsAdmin(rawToken)){
+            areaService.modifyArea(area);
+        }
+        else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"",new InvalidTokenException("meow"));
     }
 
     @DeleteMapping(path="{id}")
-    public void deleteArea(@PathVariable("id") int id) throws AreaDoesNotExistsException {
-        areaService.deleteAreaById(id);
+    public void deleteArea(@RequestHeader("authorization") String rawToken,@PathVariable("id") int id) throws AreaDoesNotExistsException, hu.otemplom.karbantarto.model.Exceptions.User.InvalidIdException, InvalidRoleException, InvalidFullNameException, InvalidUsernameException, InvalidTokenException {
+        if(authenticator.userIsAdmin(rawToken)){
+            areaService.deleteAreaById(id);
+        }
+        else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"",new InvalidTokenException("meow"));
+
     }
     @GetMapping(path="{id}")
-    public Area getAreaById(@PathVariable("id")int id) throws AreaDoesNotExistsException {
-        return areaService.getAreaById(id);
+    public Area getAreaById(@RequestHeader("authorization") String rawToken,@PathVariable("id")int id) throws AreaDoesNotExistsException, hu.otemplom.karbantarto.model.Exceptions.User.InvalidIdException, InvalidRoleException, InvalidFullNameException, InvalidUsernameException, InvalidTokenException {
+        if(authenticator.userIsAdmin(rawToken)){
+            return areaService.getAreaById(id);
+        }
+        else throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,"",new InvalidTokenException("meow"));
+
     }
     @GetMapping(path="userId={userid}")
     public Collection<Area> getAreasByUserId(@PathVariable("userid")int id){
